#24 update : daily 타입 변경

+++ data/daily/api/src/main/kotlin/com/titi/data/daily/api/model/DailyRepositoryModel.kt
@@ -4,8 +4,8 @@ data class DailyRepositoryModel(
     val id: Long,
     val status: String?,
     val day: String,
-    val timeline: List<Int>,
-    val maxTime: Int,
+    val timeline: List<Long>,
+    val maxTime: Long,
     val tasks: Map<String, Long>?,
     val taskHistories: Map<String, List<TaskHistoryRepositoryModel>>?
 )
