+++ data/daily/impl/src/main/kotlin/com/titi/data/daily/impl/local/model/DailyEntity.kt
@@ -10,8 +10,8 @@ internal data class DailyEntity(
     val id: Long,
     val status: String?,
     val day: String,
-    val timeline: List<Int>,
-    val maxTime: Int,
+    val timeline: List<Long>,
+    val maxTime: Long,
     val tasks: Map<String, Long>?,
     val taskHistories: Map<String, List<TaskHistoryEntity>>?
 )
